move post estimation out of Feols class
