+++ tests/test_feols_feglm_internally.py
@@ -48,7 +48,9 @@ def test_ols_vs_gaussian_glm(fml, inference, dropna):
     check_absolute_diff(
         fit_ols._solve_weights[0:5], fit_gaussian._solve_weights[0:5], tol=1e-10
     )
-    check_absolute_diff(fit_ols._u_hat_wls[0:5], fit_gaussian._u_hat_wls[0:5], tol=1e-10)
+    check_absolute_diff(
+        fit_ols._u_hat_wls[0:5], fit_gaussian._u_hat_wls[0:5], tol=1e-10
+    )
     check_absolute_diff(fit_ols._scores[0, :], fit_gaussian._scores[0, :], tol=1e-10)
 
     if inference == "iid":
